Generate validation schema for LND
