+++ lib/validate.js
@@ -9,7 +9,6 @@ import { NOSTR_MAX_RELAY_NUM, NOSTR_PUBKEY_BECH32, NOSTR_PUBKEY_HEX } from './no
 import { msatsToSats, numWithUnits, abbrNum, ensureB64, B64_URL_REGEX } from './format'
 import * as usersFragments from '@/fragments/users'
 import * as subsFragments from '@/fragments/subs'
-import { isInvoicableMacaroon, isInvoiceMacaroon } from './macaroon'
 import { datePivot } from './time'
 import { decodeRune } from '@/lib/cln'
 import bip39Words from './bip39-words'
@@ -156,7 +155,7 @@ const lightningAddressValidator = process.env.NODE_ENV === 'development'
     'address is no good')
   : string().email('address is no good')
 
-const hexOrBase64Validator = string().test({
+export const hexOrBase64Validator = string().test({
   name: 'hex-or-base64',
   message: 'invalid encoding',
   test: (val) => {
@@ -311,20 +310,7 @@ export function lnAddrAutowithdrawSchema ({ me } = {}) {
       test: addr => !addr.endsWith('@stacker.news'),
       message: 'automated withdrawals must be external'
     }),
-    ...autowithdrawSchemaMembers({ me })
-  })
-}
-
-export function LNDAutowithdrawSchema ({ me } = {}) {
-  return object({
-    socket: string().socket().required('required'),
-    macaroon: hexOrBase64Validator.required('required').test({
-      name: 'macaroon',
-      test: v => isInvoiceMacaroon(v) || isInvoicableMacaroon(v),
-      message: 'not an invoice macaroon or an invoicable macaroon'
-    }),
-    cert: hexOrBase64Validator,
-    ...autowithdrawSchemaMembers({ me })
+    ...autowithdrawSchemaMembers
   })
 }
 
@@ -350,16 +336,14 @@ export function CLNAutowithdrawSchema ({ me } = {}) {
         }
       }),
     cert: hexOrBase64Validator,
-    ...autowithdrawSchemaMembers({ me })
+    ...autowithdrawSchemaMembers
   })
 }
 
-export function autowithdrawSchemaMembers ({ me } = {}) {
-  return {
-    enabled: boolean(),
-    autoWithdrawThreshold: intValidator.required('required').min(0, 'must be at least 0').max(msatsToSats(BALANCE_LIMIT_MSATS), `must be at most ${abbrNum(msatsToSats(BALANCE_LIMIT_MSATS))}`),
-    autoWithdrawMaxFeePercent: floatValidator.required('required').min(0, 'must be at least 0').max(50, 'must not exceed 50')
-  }
+export const autowithdrawSchemaMembers = {
+  enabled: boolean(),
+  autoWithdrawThreshold: intValidator.required('required').min(0, 'must be at least 0').max(msatsToSats(BALANCE_LIMIT_MSATS), `must be at most ${abbrNum(msatsToSats(BALANCE_LIMIT_MSATS))}`),
+  autoWithdrawMaxFeePercent: floatValidator.required('required').min(0, 'must be at least 0').max(50, 'must not exceed 50')
 }
 
 export function bountySchema (args) {
