+++ pages/settings/wallets/[wallet].js
@@ -10,10 +10,7 @@ import Info from '@/components/info'
 import Text from '@/components/text'
 import { AutowithdrawSettings } from '@/components/autowithdraw-shared'
 import dynamic from 'next/dynamic'
-import { array, object, string } from 'yup'
-import { autowithdrawSchemaMembers } from '@/lib/validate'
-import { useMe } from '@/components/me'
-import { TOR_REGEXP } from '@/lib/url'
+import { generateSchema } from '@/lib/wallet'
 
 const WalletButtonBar = dynamic(() => import('@/components/wallet-buttonbar.js'), { ssr: false })
 
@@ -24,7 +21,6 @@ export default function WalletSettings () {
   const router = useRouter()
   const { wallet: name } = router.query
   const wallet = useWallet(name)
-  const me = useMe()
 
   const initial = wallet.fields.reduce((acc, field) => {
     // We still need to run over all wallet fields via reduce
@@ -38,7 +34,7 @@ export default function WalletSettings () {
     }
   }, wallet.config)
 
-  const schema = generateSchema(wallet, { me })
+  const schema = generateSchema(wallet)
 
   return (
     <CenterLayout>
@@ -136,96 +132,4 @@ function WalletFields ({ wallet: { config, fields } }) {
     }
     return null
   })
-}
-
-function generateSchema (wallet, { me }) {
-  if (wallet.schema) return wallet.schema
-
-  const fieldValidator = (field) => {
-    if (!field.validate) {
-      // default validation
-      let validator = string()
-      if (!field.optional) validator = validator.required('required')
-      return validator
-    }
-
-    if (field.validate.schema) {
-      // complex validation
-      return field.validate.schema
-    }
-
-    const { type: validationType, words, min, max } = field.validate
-
-    let validator
-
-    const stringTypes = ['url', 'string']
-
-    if (stringTypes.includes(validationType)) {
-      validator = string()
-
-      if (field.validate.length) {
-        validator = validator.length(field.validate.length)
-      }
-    }
-
-    if (validationType === 'url') {
-      validator = process.env.NODE_ENV === 'development'
-        ? validator
-          .or([string().matches(/^(http:\/\/)?localhost:\d+$/), string().url()], 'invalid url')
-        : validator
-          .url()
-          .test(async (url, context) => {
-            if (field.validate.torAllowed && TOR_REGEXP.test(url)) {
-              // allow HTTP and HTTPS over Tor
-              if (!/^https?:\/\//.test(url)) {
-                return context.createError({ message: 'http or https required' })
-              }
-              return true
-            }
-            try {
-              // force HTTPS over clearnet
-              await string().https().validate(url)
-            } catch (err) {
-              return context.createError({ message: err.message })
-            }
-            return true
-          })
-    }
-
-    if (words) {
-      validator = array()
-        .transform(function (value, originalValue) {
-          if (this.isType(value) && value !== null) {
-            return value
-          }
-          return originalValue ? originalValue.trim().split(/[\s]+/) : []
-        })
-        .test(async (values, context) => {
-          for (const v of values) {
-            try {
-              await string().oneOf(words).validate(v)
-            } catch {
-              return context.createError({ message: `'${v}' is not a valid ${field.label} word` })
-            }
-          }
-          return true
-        })
-    }
-
-    if (min !== undefined) validator = validator.min(min)
-    if (max !== undefined) validator = validator.max(max)
-
-    if (!field.optional) validator = validator.required('required')
-
-    return validator
-  }
-
-  return object(
-    wallet.fields.reduce((acc, field) => {
-      return {
-        ...acc,
-        [field.name]: fieldValidator(field)
-      }
-    }, wallet.walletType ? autowithdrawSchemaMembers({ me }) : {})
-  )
 }
