+++ wallets/lnd/index.js
@@ -1,4 +1,4 @@
-import { LNDAutowithdrawSchema } from '@/lib/validate'
+import { isInvoicableMacaroon, isInvoiceMacaroon } from '@/lib/macaroon'
 
 export const name = 'lnd'
 
@@ -9,7 +9,10 @@ export const fields = [
     type: 'text',
     placeholder: '55.5.555.55:10001',
     hint: 'tor or clearnet',
-    clear: true
+    clear: true,
+    validate: {
+      type: 'socket'
+    }
   },
   {
     name: 'macaroon',
@@ -21,7 +24,12 @@ export const fields = [
     type: 'text',
     placeholder: 'AgEDbG5kAlgDChCn7YgfWX7uTkQQgXZ2uahNEgEwGhYKB2FkZHJlc3MSBHJlYWQSBXdyaXRlGhcKCGludm9pY2VzEgRyZWFkEgV3cml0ZRoPCgdvbmNoYWluEgRyZWFkAAAGIJkMBrrDV0npU90JV0TGNJPrqUD8m2QYoTDjolaL6eBs',
     hint: 'hex or base64 encoded',
-    clear: true
+    clear: true,
+    validate: {
+      type: 'hexOrBase64',
+      test: v => isInvoiceMacaroon(v) || isInvoicableMacaroon(v),
+      message: 'not an invoice macaroon or an invoicable macaroon'
+    }
   },
   {
     name: 'cert',
@@ -30,7 +38,10 @@ export const fields = [
     placeholder: 'LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNNVENDQWRpZ0F3SUJBZ0lRSHBFdFdrcGJwZHV4RVF2eVBPc3NWVEFLQmdncWhrak9QUVFEQWpBdk1SOHcKSFFZRFZRUUtFeFpzYm1RZ1lYVjBiMmRsYm1WeVlYUmxaQ0JqWlhKME1Rd3dDZ1lEVlFRREV3TmliMkl3SGhjTgpNalF3TVRBM01qQXhORE0wV2hjTk1qVXdNekF6TWpBeE5ETTBXakF2TVI4d0hRWURWUVFLRXhac2JtUWdZWFYwCmIyZGxibVZ5WVhSbFpDQmpaWEowTVF3d0NnWURWUVFERXdOaWIySXdXVEFUQmdjcWhrak9QUUlCQmdncWhrak8KUFFNQkJ3TkNBQVJUS3NMVk5oZnhqb1FLVDlkVVdDbzUzSmQwTnBuL1BtYi9LUE02M1JxbU52dFYvdFk4NjJJZwpSbE41cmNHRnBEajhUeFc2OVhIK0pTcHpjWDdlN3N0Um80SFZNSUhTTUE0R0ExVWREd0VCL3dRRUF3SUNwREFUCkJnTlZIU1VFRERBS0JnZ3JCZ0VGQlFjREFUQVBCZ05WSFJNQkFmOEVCVEFEQVFIL01CMEdBMVVkRGdRV0JCVDAKMnh3V25GeHRUNzI0MWxwZlNoNm9FWi9UMWpCN0JnTlZIUkVFZERCeWdnTmliMktDQ1d4dlkyRnNhRzl6ZElJRApZbTlpZ2d4d2IyeGhjaTF1TVMxaWIyS0NGR2h2YzNRdVpHOWphMlZ5TG1sdWRHVnlibUZzZ2dSMWJtbDRnZ3AxCmJtbDRjR0ZqYTJWMGdnZGlkV1pqYjI1dWh3Ui9BQUFCaHhBQUFBQUFBQUFBQUFBQUFBQUFBQUFCaHdTc0VnQUQKTUFvR0NDcUdTTTQ5QkFNQ0EwY0FNRVFDSUEwUTlkRXdoNXpPRnpwL3hYeHNpemh5SkxNVG5yazU1VWx1NHJPRwo4WW52QWlBVGt4U3p3Y3hZZnFscGx0UlNIbmd0NUJFcDBzcXlHL05nenBzb2pmMGNqQT09Ci0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0K',
     optional: 'optional if from [CA](https://en.wikipedia.org/wiki/Certificate_authority) (e.g. voltage)',
     hint: 'hex or base64 encoded',
-    clear: true
+    clear: true,
+    validate: {
+      type: 'hexOrBase64'
+    }
   }
 ]
 
@@ -39,8 +50,6 @@ export const card = {
   subtitle: 'autowithdraw to your Lightning Labs node',
   badges: ['receive only', 'non-custodial']
 }
-
-export const schema = LNDAutowithdrawSchema
 
 export const walletType = 'LND'
